fixup! Remove/rewrite unused redux actions
